replace {date}

+++ _ttt/genride.py
@@ -13,25 +13,24 @@ if __name__ == "__main__":
         date = f'{datetime.strftime(current_date, "%Y-%m-%d")}'
         filename = f'{date}.md'
 
-        with open(filename, 'w') as f:
-        # if path.isfile(filename):
-        #     print(filename, i, 'already exists')
-        # else:
-        #     with open(filename, 'w') as f:
-        #         template = f'''---
-        #             route: ??? # (where did you go / what way)
-        #             swim: ??? # (did you swim? yes/no)
-        #             venue: # (where you ended up or ate)
-        #             ---
+        if path.isfile(filename):
+            print(filename, i, 'already exists')
+        else:
+            with open(filename, 'w') as f:
+                template = f'''---
+                    route: ??? # (where did you go / what way)
+                    swim: ??? # (did you swim? yes/no)
+                    venue: # (where you ended up or ate)
+                    ---
 
-        #             <!-- content goes here, uses markdown -->
+                    <!-- content goes here, uses markdown -->
 
-        #             <!-- images will automatically be shown, if put in images/ttt/. must match the date of the ride, in format YYYY-MM-DD. can be jpg or png -->
+                    <!-- images will automatically be shown, if put in images/ttt/. must match the date of the ride, in format YYYY-MM-DD. can be jpg or png -->
 
-        #             ![](../images/ttt/{date}.png)
-        #             ![](../images/ttt/{date}.jpg)
-        #         '''
-        #         f.write(template) 
-        #     print(f'created {filename} ride number {i}')
+                    ![](../images/ttt/{date}.png)
+                    ![](../images/ttt/{date}.jpg)
+                '''
+                f.write(template) 
+            print(f'created {filename} ride number {i}')
         current_date += week
         i += 1
